Added a BulkFileUploadIndexModel and removed the BulkFileUploadModel property from the OdsInstanceSettingsModel. Moved the BulkUpload models to a separate folder under Viewmodels.
