+++ Application/EdFi.Ods.AdminApp.Web/Models/ViewModels/OdsInstanceSettings/OdsInstanceSettingsModel.cs
@@ -9,6 +9,5 @@ namespace EdFi.Ods.AdminApp.Web.Models.ViewModels.OdsInstanceSettings
     {
         public OdsInstanceSetupCompletedModel ProductionSetupCompletedModel { get; set; }
         public LogSettingsModel LogSettingsModel { get; set; }
-        public BulkFileUploadModel BulkFileUploadModel { get; set; }
     }
 }
